Fixed reaching over logical sides bug
*Config exp coefficient numeric is now a capability accessed value requiring a PlayerEntity instance.

Previously, the expCoeff method on the server would use the server config values (like intended) but if called on the client it would use the client config values, which would result in the client side exp/level displaying inconsistent values with the actual (server) values.

Now the server gets the server side config values and sets them in the capability, and when a sync happens it would sync that value to client.

-Removed Util#expCoeff
+Added IPlayerAttributes#expCoeff