Bug fixes
*Fixed bug with button causing crash when FTB Library is installed
*Fixed bug with Artifacts causing world gen to be unresponsive and crash the game:

+Implemented (jar-in-jar) a new attribute "step-height" that Artifacts uses which was causing the above issue

Found weird interaction with artifact item "steadfast spikes", that sets Knockback Resistance to it's max value when equipped, so that when adding Knockback Resistance using PlayerEx, none is added because it's already at max; the issue is when steadfast spikes are unequipped, the bonus knockback resistance that should be there from Constitution is not present. Not sure if this one is fixable, but is not too serious since you can just unequip your items before skilling Constitution.