Language Fixes
*Made all attributes/combat screen display text to use dedicated translation keys.
+Added config options for gui text scale x/y:
This allows the display text on the attributes/combat screen to be "squished"/resized if needed, to accommodate longer translations, as well as allowing a cleaner feel for some.
+Added config option to enable/disable a tooltip fix:
This has been bugging me for a while; the item tooltip that displays attack damage/speed doesn't show the true value, only the base attribute value + the item modifier. This fixes (or will do, working on it) this, and makes the value include the extra attack damage/speed provided by levelling those attributes.

+Added relevant lang.

Still need to actually implement the tooltip fix, but it's late now and I'm sleepy.